Use only one LICENSE in the project
The dist.ini file defined the project license to be Perl_5, while our
LICENSE file in the repository root was for MIT. This was inconsistent
and also caused an error when running `dzil build`.

Now we have a Perl 5 compatible LICENSE file, and stop generating it
with Dist::Zilla::Plugin::License.